Updated the Extraction code to get SEDs, to allow for UDS and COSMOS field SEDs to be extracted
Added the ability to use the UDS and COSMOS field to extract their SEDs. This will allow the entire dataset to be explored in some detail at some point. The next step is that the SED processing code needs to be updated to match, allowing the names to accurately represent the new SED name