master dataframe for handling analysis
created a master dataframe that can be used in the analysis, we can explore how much the change in uv and vj colours affect the colour space. Ultimately we would like to see if we can create an uncertainity value for exploring how values greater than this uncertainity will affect colour space.